+++ src/noise.rs
@@ -11,14 +11,6 @@ use crate::{
 
 /// Provides modifier methods for noise types.
 pub trait Noise {
-    /// Samples the noise with seed 0.
-    fn sample<const DIM: usize, Point>(&self, point: Point) -> f32
-    where
-        Self: Sized + Sample<DIM, Point>,
-    {
-        self.sample_with_seed(point, 0)
-    }
-
     /// Samples the noise in 2D.
     fn sample2<Point>(&self, point: Point) -> f32
     where